combine maybeFetch with custom error & retries create forceFetch function
